argc: refactor ast visitors
